A server between Qualisys and another PC
